requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.piptools --output-file requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via
    #   httpx
    #   openai
    #   starlette
    #   watchfiles
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.3
    # via requests
click==8.2.1
    # via uvicorn
deprecation==2.1.0
    # via
    #   postgrest
    #   storage3
distro==1.9.0
    # via openai
fastapi==0.116.1
    # via -r requirements.piptools
gotrue==2.12.3
    # via supabase
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
h2==4.2.0
    # via httpx
hpack==4.1.0
    # via h2
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   gotrue
    #   openai
    #   postgrest
    #   storage3
    #   supabase
    #   supafunc
hyperframe==6.1.0
    # via h2
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
jiter==0.10.0
    # via openai
numpy==2.3.2
    # via
    #   -r requirements.piptools
    #   pandas
openai==1.99.5
    # via -r requirements.piptools
packaging==25.0
    # via deprecation
pandas==2.3.1
    # via -r requirements.piptools
postgrest==1.1.1
    # via supabase
pydantic==2.11.7
    # via
    #   -r requirements.piptools
    #   fastapi
    #   gotrue
    #   openai
    #   postgrest
    #   realtime
pydantic-core==2.33.2
    # via pydantic
pyjwt==2.10.1
    # via gotrue
python-dateutil==2.9.0.post0
    # via
    #   pandas
    #   storage3
python-dotenv==1.1.1
    # via
    #   -r requirements.piptools
    #   uvicorn
python-multipart==0.0.20
    # via -r requirements.piptools
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via uvicorn
realtime==2.7.0
    # via supabase
regex==2025.7.34
    # via tiktoken
requests==2.32.4
    # via
    #   -r requirements.piptools
    #   tiktoken
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   openai
starlette==0.47.2
    # via fastapi
storage3==0.12.1
    # via supabase
strenum==0.4.15
    # via supafunc
supabase==2.18.0
    # via -r requirements.piptools
supafunc==0.10.1
    # via supabase
tiktoken==0.11.0
    # via -r requirements.piptools
tqdm==4.67.1
    # via openai
typing-extensions==4.14.1
    # via
    #   anyio
    #   fastapi
    #   openai
    #   pydantic
    #   pydantic-core
    #   realtime
    #   starlette
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via requests
uvicorn==0.35.0
    # via -r requirements.piptools
uvloop==0.21.0
    # via uvicorn
watchfiles==1.1.0
    # via uvicorn
websockets==15.0.1
    # via
    #   realtime
    #   uvicorn
